Extract ComplexityWalker into separate file
